+++ src/Led/MP_Led.h
@@ -8,7 +8,7 @@ class MP_Led
   public:
 	virtual void on() const = 0;
 	virtual void off() const = 0;
-	virtual void blink(uint32_t frequency) const = 0 ;
+	virtual void blink(uint16_t frequency) const = 0 ;
 	virtual void dim(uint8_t percentage) const = 0 ;
 };
